Create `math` module for math types and impls
`f32` types are in `math/single.rs`.

`f64` types are in `math/double.rs`.

`math/mod.rs` reimports the correct types depending on the features.
